WIP: Working version for CSX
