+++ nsls2tools/csx1/startup/__init__.py
@@ -1,5 +1,4 @@
 from .startup import *
-from ..ophyd.devices import *
 from .accelerator import *
 from .optics import *
 from .endstation import *
