+++ nsls2tools/csx1/startup/endstation.py
@@ -2,6 +2,10 @@ from ophyd import (PVPositioner, EpicsMotor, EpicsSignal, EpicsSignalRO,
                    PVPositionerPC, Device)
 from ophyd import Component as Cpt
 
+from ..ophyd.devices import (SamplePosVirtualMotor, Cryoangle,
+                             Nanopositioner, Lakeshore336,
+                             VIMeter)
+
 
 # Diffo angles
